замена long long на int64_t

+++ hw-02-lucky-tickets/Counter.h
@@ -4,7 +4,7 @@
 class Counter {
 public:
 	/* Что-то считает */
-	virtual long long count(int n) const = 0LL;
+	virtual int64_t count(int n) const = 0LL;
 
 	virtual ~Counter() = default;
 };
