+++ hw-02-lucky-tickets/TestParam.cpp
@@ -1,10 +1,10 @@
 #include "TestParam.h"
 
-TestParam::TestParam(const std::string& name, int in, long long out) :
+TestParam::TestParam(const std::string& name, int in, int64_t out) :
 	name{ name }, in{ in }, out{ out } {
 };
 
 // getters
 std::string TestParam::getName() const { return name; }
 int TestParam::getIn() const { return in; }
-long long TestParam::getOut() const { return out; }
+int64_t TestParam::getOut() const { return out; }
